Moved testing suite elsewhere.
Updated coverage setup to ensure better coverage and representation.
