+++ Python/.coveragerc
@@ -3,18 +3,20 @@ omit =
 	*/python?.?.?/*	
 	*__init__*
 	*venv*
-
+	*/tests/*
+	*/Coverage/*
 [report]
 omit = 
 	*/python?.?.?/*	
 	*__init__*
 	*venv*
+	*/tests/*
+	*/Coverage/*
 exclude_lines =
     pragma: no cover
     def __repr__
     if self.debug:
     if settings.DEBUG
-    raise AssertionError
     raise NotImplementedError
     if 0:
     if __name__ == .__main__.:
